lab 08 solution

+++ 08-traits/src/main/scala/Shape.scala
@@ -1,6 +1,6 @@
 trait Shape {
 
-  def area(): Int
+  def area(): Double
 
   // another one
   // def circumference() : Int
